+++ clifford/__init__.py
@@ -62,7 +62,6 @@ Miscellaneous functions
 """
 
 # Standard library imports.
-from functools import reduce
 import os
 import itertools
 from typing import List, Tuple, Set, Container, Dict, Optional
@@ -310,170 +309,9 @@ def val_get_right_gmt_matrix(mt: sparse.COO, x):
 from ._layout import Layout  # noqa: E402
 from ._multivector import MultiVector  # noqa: E402
 from ._conformal_layout import ConformalLayout  # noqa: E402
-from ._mvarray import MVArray  # noqa: E402
-
-
-def array(obj):
-    '''
-    an array method like numpy.array(), but returns a MVArray
-
-    Parameters
-    -------------
-    obj : MultiVector, list
-        a MV or a list of MV's
-
-    Examples
-    ----------
-    >>>import clifford as cf
-    >>>from clifford import g3
-    >>>import numpy as np
-    >>>np.random.rand(10)*cf.array(g3.e12)
-    '''
-    if isinstance(obj, MultiVector):
-        # they passed a single MV so make a list of it.
-        return MVArray([obj])
-    else:
-        return MVArray(obj)
-
-
-class Frame(MVArray):
-    '''
-    A frame of vectors
-    '''
-    def __new__(cls, input_array):
-        if not np.all([k.grades() == {1} for k in input_array]):
-            raise TypeError('Frames must be made from vectors')
-
-        obj = MVArray.__new__(cls, input_array)
-        return obj
-
-    def __array_finalize__(self, obj):
-        if obj is None:
-            return
-
-    @property
-    def En(self):
-        '''
-        Volume element for this frame
-
-        En = e1^e2^...^en
-        '''
-        return reduce(op, self)
-
-    @property
-    def inv(self):
-        '''
-        The inverse frame of self
-
-        Returns
-        ---------
-        inv : `clifford.Frame`
-        '''
-
-        En = self.En
-        # see D&L sec 4.3
-        vectors = [
-            (-1)**(k)*reduce(op, np.hstack([self[:k], self[k+1:]]))*En.inv()
-            for k in range(len(self))]
-
-        return Frame(vectors)
-
-    def is_innermorphic_to(self, other, eps=None):
-        '''
-        Is this frame `innermorhpic` to  other?
-
-        *innermorphic* means both frames share the same inner-product
-        between corresponding vectors. This implies that the two frames
-        are related by an orthogonal transform
-
-        Parameters
-        ------------
-        other : `clifford.Frame`
-            the other frame
-
-        Returns
-        ----------
-        value : bool
-
-        '''
-        # make iterable `pairs` of all index combos, without repeat
-        pairs = list(itertools.combinations(range(len(self)), 2))
-        a, b = self, other
-        if eps is None:
-            eps = _eps
-
-        return np.array([
-            float((b[m]|b[n]) - (a[m]|a[n])) < eps
-            for m, n in pairs
-        ]).all()
-
-
-class BladeMap(object):
-    '''
-    A Map Relating Blades in two different algebras
-
-    Examples
-    -----------
-    >>> from clifford import Cl
-    >>> # Dirac Algebra  `D`
-    >>> D, D_blades = Cl(1, 3, firstIdx=0, names='d')
-    >>> locals().update(D_blades)
-
-    >>> # Pauli Algebra  `P`
-    >>> P, P_blades = Cl(3, names='p')
-    >>> locals().update(P_blades)
-    >>> sta_split = BladeMap([(d01, p1),
-                              (d02, p2),
-                              (d03, p3),
-                              (d12, p12),
-                              (d23, p23),
-                              (d13, p13)])
-
-    '''
-    def __init__(self, blades_map, map_scalars=True):
-        self.blades_map = blades_map
-
-        if map_scalars:
-            # make scalars in each algebra map
-            s1 = self.b1[0]._newMV(dtype=int)+1
-            s2 = self.b2[0]._newMV(dtype=int)+1
-            self.blades_map = [(s1, s2)] + self.blades_map
-
-    @property
-    def b1(self):
-        return [k[0] for k in self.blades_map]
-
-    @property
-    def b2(self):
-        return [k[1] for k in self.blades_map]
-
-    @property
-    def layout1(self):
-        return self.b1[0].layout
-
-    @property
-    def layout2(self):
-        return self.b2[0].layout
-
-    def __call__(self, A):
-        '''map an MV `A` according to blade_map'''
-
-        # determine direction of map
-        if A.layout == self.layout1:
-            from_b = self.b1
-            to_b = self.b2
-
-        elif A.layout == self.layout2:
-            from_b = self.b2
-            to_b = self.b1
-        else:
-            raise ValueError('A doesnt belong to either Algebra in this Map')
-
-        # create empty MV, and map values
-        B = to_b[0]._newMV(dtype=int)
-        for from_obj, to_obj in zip(from_b, to_b):
-            B += (sum(A.value*from_obj.value)*to_obj)
-        return B
+from ._mvarray import MVArray, array  # noqa: F401
+from ._frame import Frame  # noqa: F401
+from ._blademap import BladeMap  # noqa: F401
 
 
 # copied from the itertools docs
